Add: Adiciona possibilidade de ter um banco de dados.

+++ hooks/post_gen_project.py
@@ -15,21 +15,27 @@ def criar_gitignore():
     os.remove('.gitignore_project')
 
 
-def remover_arquivos_extras(proj_tipo, use_jupyter):
+def remover_arquivos_extras(proj_tipo, use_jupyter, database):
     if (use_jupyter == "No"):
         os.remove('Modelling_{{cookiecutter.project_name}}.ipynb')
         os.remove('EDA_{{cookiecutter.project_name}}.ipynb')
-    elif (proj_tipo == "EDA"):
-        os.remove('Modelling_{{cookiecutter.project_name}}.ipynb')
+    if (proj_tipo == "EDA"):
+        if (use_jupyter != "No"):
+            os.remove('Modelling_{{cookiecutter.project_name}}.ipynb')
         shutil.rmtree('./models')
         shutil.rmtree('./models_results')
+    if (database == "No"):
+        os.remove('docker-compose.yml')
+        os.remove('.database.env')
+        shutil.rmtree('./sql')
 
 
 def main():
     proj_tipo = "{{cookiecutter.project_type}}"
     use_jupyter = "{{cookiecutter.use_jupyter}}"
+    database = "{{cookiecutter.database}}"
     criar_gitignore()
-    remover_arquivos_extras(proj_tipo, use_jupyter)
+    remover_arquivos_extras(proj_tipo, use_jupyter, database)
 
 
 if __name__ == '__main__':
